Added loading component (for multiple deletes). Table refresh function added after update, delete, create operations.

+++ src/components/Modals/DeleteModal.jsx
@@ -1,5 +1,5 @@
 import { Modal,Button,Box } from '@mui/material'
-import React,{ useContext,useReducer }  from 'react'
+import React,{ useContext,useReducer,useState,useEffect }  from 'react'
 import { deleteUser } from '../../services/api';
 import { TableContext } from '../../context/TableContext';
 import { ModalContext } from '../../context/ModalContext';
@@ -10,23 +10,23 @@ import { styles } from '../CustomStyles';
   
 
 function DeleteModal() {
+  
     const {refreshDataTable} = useContext(TableContext)
     const {state,dispatch} = useContext(ModalContext)
     
-
-    const handleDelete = () => {
-        deleteUser(state.selectedUserId)
-        refreshDataTable(state.selectedUserId)
-        dispatch({type: 'SET_SELECTED_USER_ID',payload: null})
-        dispatch({type: 'CLOSE_DELETE_MODAL'})
+    const handleDelete = async () => {
+        if(state.selectedUserId !== undefined){
+            await deleteUser(state.selectedUserId)
+            refreshDataTable(state.selectedUserId)
+            handleCloseDeleteModal()
+     
+        }
     }
 
     const handleCloseDeleteModal = () => {
+        dispatch({type: 'SET_SELECTED_USER_ID',payload: null})
         dispatch({type: 'CLOSE_DELETE_MODAL'})
     };
-
-   
-
 
   return (
     <Modal 
@@ -36,7 +36,7 @@ function DeleteModal() {
         <Box sx={styles.deleteModalBoxStyle}>
             <span className='delete-confirm'>Are you sure want to delete this user?</span>
             
-            <Box >
+            <Box sx={{marginTop: '30px'}}>
                 <Button color="error" variant='contained' onClick={handleDelete}>Delete</Button>
                 <Button variant="outlined" onClick={handleCloseDeleteModal} sx={{marginLeft: '20px'}}>Cancel</Button>
             </Box>
